Move accumulators tests to their own file
